requirements.txt
Django==3.2.13
django-cors-headers==3.11.0
django-filebrowser==4.0.1
django-grappelli==3.0.2
django-nested-admin==3.4.0
djangorestframework==3.13.1
mysqlclient==2.1.0
django-import-export==2.6.1